#29165 If interpreter is not found, the "python not found" is displayed
This change will display a new "python not found" UI whenever the python
interpreter file is not found. Previously, there would be a "setup on
new operating system" UI displayed no value had been defined in the
interpreter was found (which corresponded to an uninitialized OS). We
are now seeding all OSes with default interpreter paths to help make
things work out of the box, meaning that the case where there is a blank
interpreter file is much less common. Also added more logging around
this to help support and debugging.

The copyright message in the about dialog now says 2015

Fixed issues with font sizes in UIs being different on windows, linux
and macosx.

Closes #18