#31384 Add tabs for apps that register panels
When an application used by tk-desktop registers a panel, this panel is
now initialized and its widget is added as a tab in the desktop
application view. A button is also added to the header of the
application to allow switching tabs.

For the moment, that `Apps` tab is a "hard-coded" tab added
automatically. Panels can register by passing a "position" value. The
tabs are then sorted (ascending order), with the ones lacking a
"position" setting placed at the end (after the ones with specified
positions). `Apps` has a default 0 position.

UI details:
- removed `Apps` button. It is now automatically generated, like any
  other tab
- renamed the `stack` QStackWidget to `apps_tab`
- added a QStackWidget `tab_view` to act as a parent of tab widgets
- modification of the css to better show the state of tab header buttons

Closes #25.