Apps tab must be specified in run_at_startup
Instead of blindly adding the "Apps" tab as the first tab, we look for a
selector that matches Desktop's instance name as well as an "Apps"
command (or a wildcard empty string).
In order to do so, we change to logic: find the fake selector from the
list of selectors in the config, strip the fake selector off (as
get_matching_commands will show warnings otherwise), get the matching
commands, figure where we should place a fake "Apps" callback in the
list of matching commands and call the command callbacks.
Figuring out where to place the fake "Apps" callback is tricky because
we can have corner cases:
- selectors that fail to match commands
- selectors that match multiple commands
Hopefully the comments properly address this complexity.