For #29128: Don't use the event log for tracking recent activity
Previously we used the event log to help track recent app launches and recently accessed projects per user. This was causing performance issues due to the additional queries required to calculate this information (especially on the event log which can be very slow).

With this change, we now track recent app launches using the settings module in the Shotgun Utils Framework. These are tracked for the user on a per-project level and are unique to each machine the user is using. Recently accessed Projects are now solely tracked by the Project.last_accessed_by_currrent_user field on the Project in Shotgun. This information is updated whenever the Project is accessed via SG Desktop as well.