For #39630, UI to display when no commands are found for a project. (#47)
If no executable paths are found on the local filesystem corresponding to registered Software entities for a site, display a UI that shows the Software entity icons, a message that nothing was found to launch, and a link to a help page describing how to setup Software entities.  The link is  actually pointing to the documentation for tk-multi-launchapp and will need to be updated once the Software documentation is complete. Closes #47.