#38927, User menu item to launch classic Project Setup Wizard (#50)
Adds an 'Advanced project setup...' action to the User menu for projects that bootstrap Toolkit directly using a predefined configuration (tk-config-basic). This menu item is not available on the Projects page or for projects that have one or more existing pipeline configurations in Shotgun. 

Also implements a place-holder QMessageBox information popup dialog that links to the help documentation for the Toolkit Project setup wizard. This help dialog should popup only once for each user that has permission to setup a Toolkit project.