#40549, Support 'group' and 'group_default' command properties (#57)
If no collapse rules have been determined for a command, check for input group and group_default properties. If present, these will be treated the same as collapse rules. The current implementation does not actually support the group_default property, as this is determined at a much lower level by the project_commands_model.