For #43269: Fix issue with single configuration pipeline (#90)
Moved the creation of the menu separator to the 'ProjectMenu.clear()' which was also renamed 'rese()' for clarity. On switching back and forth between a multi-configuration pipeline and a single configuration pipeline, a call to the 'ProjectMenu.clear()'  method would be made. The later clears the QT menu and the '_pipeline_configuration_separator' instance member would then reference a deleted QT object.