SG-15442 Fixes a bug with closing a Python process (#103)
A regression in the last release introduced a bug where it errored when shutting down the project python process (ie leaving the project in SG Desktop). This was due to a typo in the method call. However, it should also now check that the logging proxy is still connected before trying to send log messages.
It also fixes a typo in a comment.