SG-17639 Fixes styling issues around docking specifically on Mac. (#113)
Fixes the arrow not always being placed in the correct position. This ensures the window geometry is updated before trying to draw the arrow.

Fixes the issue with the masking and transparency not working. Removed a setting specific to MacOS that doesn't appear to affect PySide and is recommended against in Qt5, this fixed the masking issues.

It also includes a fix that JF did for the docking animation not working in PySide2. I also implemented the fix here as I couldn't fix the other issue without fixing that.