Fixes and issue on Windows with window borders. (#118)
When pinning and then unpinning SG Desktop in windows, it would result in an old style Window border/frame for the application. This fixes that so that the styling stays the same after unpinning.

The fix doesn't feel like it gets to the bottom of the cause of the issue though. It feels like it should have worked without this fix.