SG-17640 Macos item highlighting fix (#117)
This fixes an issue where when entering a project in SG Desktop on Mac, the proxy Python process bootstrapping the project config, would steal focus when it spawned its app.

The fix here is to push the spawn proxy app to the background, and this will result in the previously active app becoming active again. So in most cases this will mean Shotgun Desktop becomes active again, however if for example you clicked on a project and then went to chrome, it would not return focus to chrome rather than stealing it, which seems like the right solution.

This however only works on Python 3, and requires the user to be using an up-to-date version of the `tk-desktop` engine in their project config.