SG-18612 fix for no keyboard input on mac (#127)
Fixes an issue where Keyboard input was ignored on any project apps.
This was because of a recent change that forced the QApplication to be created in the background. The QApplication is still created in the background when the project is bootstrapped in Desktop, however when the user selects to launch an app, we now shift it to the foreground.