SG-19710 Fixes error loading console position and size (#129)
In the seemingly rare situation where the user is running Desktop with PySide2, but the settings for console position and size are previously stored as PySide objects, and the settings manager is somehow able to load them, then this will ignore the errors that PySide2 will throw when passed a PySide object. Effectively meaning the old settings won't be loaded.