SS-18239: undocked by default (#132)
The Shotgun Desktop will now be displayed unpinned by default the first time it is launched. This will reduce the confusion of first time users who do not understand where the app goes after the splash screen disappears.

Also, when a user clicks on the X or red button of the dialog, a pop-up will be shown the very first time letting the user know the app is now in the tray. Unfortunately, this only works on Qt5 builds of Desktop, so v1.6.1 and up.

Finally, unhandled exceptions are now going to be logged.